requirements.txt
PyQt6>=6.0.0
pyserial>=3.5
requests>=2.28.0
